requirements.txt
fastapi
uvicorn
pandas
openpyxl
xlsxwriter
python-multipart
numpy
scikit-learn
python-dotenv
google-generativeai